Add `yarn check` script and fix types
Add a yarn script to typecheck the node/web projects. Unsure if there is
a better way to do this, so feel free to suggest another way if I'm just
missing something!

I denied the errors for unused parameters and locals as those seem
largely benign and I wanted to minimize the diff.

I have used the result for a bit and haven't hit any issues, but I may
have missed something.